Move tag render elements
